+++ Atlas/includes/Atlas/core/RgbSpectrum.h
@@ -213,10 +213,10 @@ namespace atlas
 		const float green_s = std::floor(green_c / pow(2.f, (exp_shared - B - N)) + 0.5f);
 		const float blue_s = std::floor(blue_c / pow(2.f, (exp_shared - B - N)) + 0.5f);
 
-		uintToBits(red_s, cw, 9, 0);
-		uintToBits(green_s, cw, 9, 9);
-		uintToBits(blue_s, cw, 9, 18);
-		uintToBits(exp_shared, cw, 5, 27);
+		uintToBits(static_cast<uint32_t>(red_s), cw, 9, 0);
+		uintToBits(static_cast<uint32_t>(green_s), cw, 9, 9);
+		uintToBits(static_cast<uint32_t>(blue_s), cw, 9, 18);
+		uintToBits(static_cast<uint32_t>(exp_shared), cw, 5, 27);
 
 		return (cw);
 	}
@@ -228,10 +228,10 @@ namespace atlas
 		const float N = 9; // N is the number of mantissa bits per component
 		const float B = 15; // B is the exponent bias
 
-		const float red_s = bitsToUint(weight, 9, 0);
-		const float green_s = bitsToUint(weight, 9, 9);
-		const float blue_s = bitsToUint(weight, 9, 18);
-		const float exp_shared = bitsToUint(weight, 5, 27);
+		const float red_s = static_cast<float>(bitsToUint(weight, 9, 0));
+		const float green_s = static_cast<float>(bitsToUint(weight, 9, 9));
+		const float blue_s = static_cast<float>(bitsToUint(weight, 9, 18));
+		const float exp_shared = static_cast<float>(bitsToUint(weight, 5, 27));
 
 		color.r = red_s * pow(2.f, exp_shared - B - N);
 		color.g = green_s * pow(2.f, exp_shared - B - N);
